Invoice total override, fix company invoices filter, improve samples.
Allow the total for an invoice to be set in frontmatter. Probably in
most cases an item with a negative amount should be used instead but
this may be necessary for invoices that do not have individual
amounts for items or other scenarios.

If a company name was fully included at the start of another company
name, e.g. 'company' in 'company2', the invoices for both companies
would be matched by the filter. Fixed by including a separator after
the company slug in the filter.

The samples have been improved with more variation to demonstrate
different ways of creating invoices as well as verifying that the
output matches expected behaviour for both the dev server and built
result.